Fix logging and mouse wheel bug
* Logging output is disabled for nomlib's integration test if SDL2
is shutdown in the first example.

* The mouse wheel was only responding to up (positive) events due to
the variable's type being unsigned.

* Change FileInterface's root path to Resources root so we can reload
documents from the non "built" / distributed file source.

* Disable SDL2SamplesTest.

+++ src/librocket/RocketSDL2SystemInterface.cpp
@@ -388,7 +388,7 @@ int RocketSDL2SystemInterface::TranslateMouseButton(Uint8 button)
     }
 }
 
-int RocketSDL2SystemInterface::TranslateMouseWheel( uint8 axis )
+int RocketSDL2SystemInterface::TranslateMouseWheel( int32 axis )
 {
   // Rocket treats a negative delta as up movement (away from the user),
   // positive as down.
